Initial commit
So I wasn't entirely sure where to place the 'thank you' form. I noticed the 'about' page was pretty blank, so I decided to place it there. (Hope that's okay!). Also because I understand this is a personally forked repo I am pushing to master and not some other branch.

Lastly, I made a few updates to several config files because I had issues with 'Prettier' getting in the way of actually debugging / writing test code.